Step 6: Move heap allocator from kernel to sys
Move Heap, SpinlockHeap, FreeBlock, AlignedSizeWithMetadata, and the
GlobalAlloc impl to sys/src/memory/heap.rs. Kernel keeps only the
#[global_allocator] static and allocated_size().
