+++ sys/src/memory/mod.rs
@@ -1,4 +1,5 @@
 pub mod address;
+pub mod heap;
 pub mod page;
 pub mod page_allocator;
